requirements_secure.txt
python-telegram-bot==20.5
yt-dlp==2023.9.24
requests==2.31.0
aiohttp==3.8.5
aiofiles==23.2.1
python-dotenv==1.0.0